+++ edx_info_pages/urls.py
@@ -1,0 +1,5 @@
+from django.urls import include, path
+
+urlpatterns = [
+    path('tinymce/', include('tinymce.urls'))
+]